sender/requirements.txt
bs4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
lxml
